Initial commit
ESU.py - Finds all subgraphs of a k size and calls canon.py for canonize the graph.
canon.py - Canonize a graph and return the list of correspondence of the graph to the canonized graph.
graphic.py - Generates graphics about the use of time, subgraph of size k founded and memory used.
preparation_time.sh - Generates some graphs with differences regarding number of nodes, edges and number k. Calls ESU.py.
transf.py - transform a graph in a another that is more easy to use in algorithms.